notifier logic fixed, added notifications for it

+++ services/notification/detailed_notification.py
@@ -73,7 +73,8 @@ class DetailedNotificationService(
                     self.model.receiver_id == user_id)
             .first()
         )
-        
+        if notification is None:
+            return
         db.delete(notification)
         db.commit()
